Only snap catchment areas to wastewater nodes
Catchment areas theoretically can be connected to all network element
types including reaches.
In practice we now force people to use nodes.

This is a workaround for situations where catchment areas were connected
to reaches instead of nodes unintentionally because priorization of
network elements was done in the wrong order.

This can be revisited in the future, if the requirement for connection
catchment areas directly to reaches re-appears.

Fix #373